changed to prevent overwriting existing spec files

+++ set_up_test_env.sh
@@ -4,13 +4,19 @@ export PYTHONPATH=$PWD
 mkdir -p $PWD/../Data_services_storage
 export DATA_SERVICES_STORAGE=$PWD/../Data_services_storage/
 
-cp $PWD/default-sources-spec.yml $PWD/../Data_services_storage/sources-spec.yml
+SOURCES_SPEC_FILE=$PWD/../Data_services_storage/sources-spec.yml
+if [ ! -f "$SOURCES_SPEC_FILE" ]; then
+    cp $PWD/default-sources-spec.yml "$SOURCES_SPEC_FILE"
+fi
 export DATA_SERVICES_SOURCES_SPEC=sources-spec.yml
 
 mkdir -p $PWD/../Data_services_graphs
 export DATA_SERVICES_GRAPHS=$PWD/../Data_services_graphs/
 
-cp $PWD/default-graph-spec.yml $PWD/../Data_services_graphs/graph-spec.yml
+GRAPH_SPEC_FILE=$PWD/../Data_services_graphs/graph-spec.yml
+if [ ! -f "$GRAPH_SPEC_FILE" ]; then
+    cp $PWD/default-graph-spec.yml $PWD/../Data_services_graphs/graph-spec.yml
+fi
 export DATA_SERVICES_GRAPH_SPEC=graph-spec.yml
 
 mkdir -p $PWD/../Data_services_storage/logs
